chore: base cleaup game manager
