+++ libs/contracts/src/game-manager.ts
@@ -1,0 +1,1 @@
+export interface IGameManager {}
