+++ libs/contracts/src/game.ts
