+++ libs/contracts/src/index.ts
@@ -1,3 +1,5 @@
 export * from './resources';
 export * from './constants';
 export * from './base-layout';
+export * from './game-manager';
+export * from './game';
