+++ libs/core/src/theme.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { ThemeProvider } from '@emotion/react';
-
-import { useLayout } from './layout';
 import { observer } from 'mobx-react-lite';
+import { useBaseLayout } from '@qspider/providers';
 
 export const Theme: React.FC = observer(({ children }) => {
-  const layout = useLayout();
+  const layout = useBaseLayout();
   return <ThemeProvider theme={layout.theme}>{children}</ThemeProvider>;
 });
