+++ libs/providers/src/index.ts
@@ -1,1 +1,2 @@
 export * from './resources';
+export * from './base-layout';
